feat: initial project setup
