+++ .gitignore
@@ -1,6 +1,9 @@
 # dependencies
 /node_modules
 
+# build
+/dist
+
 # System Files
 .DS_Store
 Thumbs.db
